与地址和用户关联的表的关联字段id都改为long

+++ src/main/java/com/sftc/web/model/Paging.java
@@ -3,15 +3,15 @@ package com.sftc.web.model;
 
 public class Paging {
 
-    private int user_id;
+    private long user_id;
     private int pageNum;
     private int pageSize;
 
-    public int getUser_id() {
+    public long getUser_id() {
         return user_id;
     }
 
-    public void setUser_id(int user_id) {
+    public void setUser_id(long user_id) {
         this.user_id = user_id;
     }
